Deduplicate, sort output from cert_database lookup
`cert_database.lookup()` uses a list for storing subdomains, but the
respose from the service can contain (many) duplicate subdomains.  In
addition, the subdomain list is not in a deterministic order, so future
diffs may be inaccurate.

This PR switches from a list to a `set` for storing the subdomains,
which automatically deduplicates them.  It also returns a sorted
version, guaranteeing a consistent order for every call.